fix: Embedded audio from MOV files with single-truth A/V sync model
- AssetCache now decodes embedded audio to RAM (like AudioSource buffers)
- Removed VideoEmbedAudioProcessor - audio flows through AudioSourceProcessor
- Fixed gain=0 issue: buffer.set(0) was removed during Phase 3, restored in Session::audioOut()
- Implemented single-truth model: audio playhead is authority, video slaves to it
- Added comprehensive diagnostics for parameter tracing (gain, sync)
- Renamed InputManager → InputBindings for hardware input abstraction
- Added VideoCache for zero-copy video swapping with cache management
- Fixed pattern triggering with proper SET_PARAM routing to audio thread
- AVSampler now uses EXTERNAL clock mode for video (driven by audio)

Audio now outputs correctly and A/V syncs perfectly with pattern sequencing.

+++ src/core/Transport.cpp
@@ -1,6 +1,6 @@
 #include "Transport.h"
 
-void Transport::update(float dt) {
+void Transport::update(double dt) {
     if (!isPlaying) return;
 
     absoluteTime += dt;
